split class files
